+++ src/main.rs
@@ -1,23 +1,48 @@
 mod api;
 mod config;
-mod display;
 mod model;
 mod portfolio;
 use api::fetch_price;
+use clap::{App, Arg};
 use config::{get_config_path, Config};
-use display::{print_sum_row, print_table_header, print_table_row};
-use model::TableRow;
-use portfolio::{process_portfolio_data, summarize_portfolio};
+use portfolio::{create_summary_row, process_portfolio_data};
 use reqwest::Client;
 use std::env;
+use tabled::settings::{
+    object::{Columns, Object, Rows},
+    Alignment, Border, Padding, Style,
+};
+use tabled::Table;
 
 #[tokio::main]
 async fn main() {
-    // Get arguments
-    let args: Vec<String> = env::args().collect();
+    let matches = App::new(env!("CARGO_PKG_NAME"))
+        .version(env!("CARGO_PKG_VERSION"))
+        .author(env!("CARGO_PKG_AUTHORS"))
+        .about(env!("CARGO_PKG_DESCRIPTION"))
+        .arg(
+            Arg::with_name("config")
+                .short('c')
+                .long("config")
+                .value_name("FILE")
+                .help("Sets a custom config file")
+                .takes_value(true),
+        )
+        .arg(
+            Arg::with_name("sort")
+                .short('s')
+                .long("sort")
+                .value_name("ORDER")
+                .help("Sets the sort order: h for hourly, w for weekly, m for monthly")
+                .takes_value(true),
+        )
+        .get_matches();
 
-    // Get the config path if any
-    let config_path = get_config_path(&args);
+    // Use the get_config_path function to get the path to the config file
+    let config_path = get_config_path(matches.value_of("config"));
+
+    // Get the sort order if provided, defaulting to 'd' for daily
+    let sort_order = matches.value_of("sort").unwrap_or("d");
 
     // Load the config file
     let config = Config::load(&config_path).unwrap_or_else(|err| {
@@ -39,138 +64,23 @@ async fn main() {
     // Make the api request and handle data
     match fetch_price(&client, &tickers_string, &config.api_key).await {
         Ok(response) => {
-            let table_rows = process_portfolio_data(&config.portfolio, &response);
+            let mut table_rows = process_portfolio_data(&config.portfolio, &response, sort_order);
 
-            print_table_header();
-            for row in &table_rows {
-                print_table_row(row);
-            }
+            let summary_row = create_summary_row(&table_rows);
+            table_rows.push(summary_row);
 
-            let (
-                total_value,
-                weighted_average_percent_change,
-                cumulative_pl,
-                cumulative_pl_percentage,
-            ) = summarize_portfolio(&table_rows);
+            let table = Table::new(&table_rows)
+                .with(Style::psql())
+                .modify(
+                    Columns::new(1..).not(Columns::first()),
+                    Padding::new(5, 1, 0, 0),
+                )
+                .modify(Columns::new(1..).not(Columns::first()), Alignment::right())
+                .modify(Rows::last(), Border::new().set_top('-').set_bottom('-'))
+                .to_string();
 
-            print_sum_row(
-                total_value,
-                weighted_average_percent_change,
-                cumulative_pl,
-                cumulative_pl_percentage,
-            );
+            println!("{}", table);
         }
         Err(err) => eprintln!("Error fetching prices: {}", err),
     }
 }
-
-// #[tokio::main]
-// async fn main() {
-//     let args: Vec<String> = env::args().collect();
-//     let config_path = get_config_path(&args);
-
-//     let config = Config::load(&config_path).unwrap_or_else(|err| {
-//         eprintln!("Error loading config: {}", err);
-//         std::process::exit(1);
-//     });
-
-//     let client = Client::new();
-
-//     let tickers: Vec<String> = config
-//         .portfolio
-//         .iter()
-//         .map(|entry| entry.ticker.clone())
-//         .collect();
-//     let tickers_string = tickers.join(",");
-
-//     match fetch_price(&client, &tickers_string, &config.api_key).await {
-//         Ok(response) => {
-//             let mut table_rows: Vec<TableRow> = Vec::new();
-
-//             for entry in &config.portfolio {
-//                 if let Some(currency) = response.data.get(&entry.ticker) {
-//                     if let Some(quote) = currency.quote.get("USD") {
-//                         let value = entry.amount * quote.price;
-
-//                         let (pl, pl_percent) = match entry.entry_price {
-//                             Some(entry_price) if entry_price > 0.0 => {
-//                                 // Perform P&L calculations only if entry_price is meaningful
-//                                 let pl = (quote.price - entry_price) * entry.amount;
-//                                 let pl_percent =
-//                                     ((quote.price - entry_price) / entry_price) * 100.0;
-//                                 (Some(pl), Some(pl_percent))
-//                             }
-//                             _ => (None, None), // Skip P&L calculations if entry_price is not meaningful or not present
-//                         };
-
-//                         table_rows.push(TableRow {
-//                             price: quote.price,
-//                             entry_price: entry.entry_price,
-//                             amount: entry.amount,
-//                             ticker: entry.ticker.clone(),
-//                             percent_change: quote.percent_change_24h,
-//                             value,
-//                             pl,
-//                             pl_percent,
-//                         });
-//                     } else {
-//                         eprintln!("No USD quote available for {}", entry.ticker);
-//                     }
-//                 } else {
-//                     eprintln!("No data available for {}", entry.ticker);
-//                 }
-//             }
-
-//             table_rows.sort_by(|a, b| {
-//                 b.percent_change
-//                     .partial_cmp(&a.percent_change)
-//                     .unwrap_or(std::cmp::Ordering::Equal)
-//             });
-//             print_table_header();
-
-//             for row in &table_rows {
-//                 print_table_row(row);
-//             }
-
-//             let total_value: f64 = table_rows.iter().map(|row| row.value).sum();
-//             let weighted_average_percent_change = if total_value > 0.0 {
-//                 table_rows
-//                     .iter()
-//                     .map(|row| {
-//                         let asset_value = row.value;
-//                         let weight = asset_value / total_value;
-//                         row.percent_change * weight
-//                     })
-//                     .sum()
-//             } else {
-//                 0.0
-//             };
-
-//             let cumulative_pl: f64 = table_rows
-//                 .iter()
-//                 .filter_map(|row| row.pl) // This filters out None values and unwraps Some values
-//                 .sum();
-
-//             let total_initial_value: f64 = table_rows
-//                 .iter()
-//                 .filter_map(|row| row.entry_price.map(|price| row.amount * price))
-//                 .sum();
-
-//             let total_current_value: f64 = table_rows.iter().map(|row| row.value).sum();
-
-//             let cumulative_pl_percentage: f64 = if total_initial_value > 0.0 {
-//                 ((total_current_value - total_initial_value) / total_initial_value) * 100.0
-//             } else {
-//                 0.0 // Avoid division by zero if total_initial_value is 0
-//             };
-
-//             print_sum_row(
-//                 total_value,
-//                 weighted_average_percent_change,
-//                 cumulative_pl,
-//                 cumulative_pl_percentage,
-//             );
-//         }
-//         Err(err) => eprintln!("Error fetching prices: {}", err),
-//     }
-// }