+++ src/portfolio.rs
@@ -5,6 +5,7 @@ use crate::model::TableRow;
 pub fn process_portfolio_data(
     portfolio: &[PortfolioEntry],
     response: &CryptoResponse,
+    sort_order: &str,
 ) -> Vec<TableRow> {
     let mut table_rows: Vec<TableRow> = Vec::new();
 
@@ -25,12 +26,14 @@ pub fn process_portfolio_data(
                         });
 
                     table_rows.push(TableRow {
-                        price: quote.price,
+                        price: Some(quote.price),
                         entry_price: entry.entry_price,
                         // Since amount is now an Option, we need to provide it directly
                         amount: Some(amount),
                         ticker: entry.ticker.clone(),
-                        percent_change: quote.percent_change_24h,
+                        hourly_percent_change: quote.percent_change_1h,
+                        daily_percent_change: quote.percent_change_24h,
+                        weekly_percent_change: quote.percent_change_7d,
                         value: Some(value),
                         pl,
                         pl_percent,
@@ -38,11 +41,13 @@ pub fn process_portfolio_data(
                 } else {
                     // Handle case where amount is None - possibly push a row with default or None values
                     table_rows.push(TableRow {
-                        price: quote.price,
+                        price: Some(quote.price),
                         entry_price: entry.entry_price,
                         amount: None,
                         ticker: entry.ticker.clone(),
-                        percent_change: quote.percent_change_24h,
+                        hourly_percent_change: quote.percent_change_1h,
+                        daily_percent_change: quote.percent_change_24h,
+                        weekly_percent_change: quote.percent_change_7d,
                         value: None,
                         pl: None,
                         pl_percent: None,
@@ -52,13 +57,38 @@ pub fn process_portfolio_data(
         }
     }
 
-    table_rows.sort_by(|a, b| {
-        b.percent_change
-            .partial_cmp(&a.percent_change)
-            .unwrap_or(std::cmp::Ordering::Equal)
-    });
+    match sort_order {
+        "h" => table_rows.sort_by(|a, b| {
+            b.hourly_percent_change
+                .partial_cmp(&a.hourly_percent_change)
+                .unwrap_or(std::cmp::Ordering::Equal)
+        }),
+        "d" => table_rows.sort_by(|a, b| {
+            b.daily_percent_change
+                .partial_cmp(&a.daily_percent_change)
+                .unwrap_or(std::cmp::Ordering::Equal)
+        }),
+        "w" => table_rows.sort_by(|a, b| {
+            b.weekly_percent_change
+                .partial_cmp(&a.weekly_percent_change)
+                .unwrap_or(std::cmp::Ordering::Equal)
+        }),
+        _ => table_rows.sort_by(|a, b| {
+            b.daily_percent_change
+                .partial_cmp(&a.daily_percent_change)
+                .unwrap_or(std::cmp::Ordering::Equal)
+        }),
+    };
 
     table_rows
+
+    // table_rows.sort_by(|a, b| {
+    //     b.daily_percent_change
+    //         .partial_cmp(&a.daily_percent_change)
+    //         .unwrap_or(std::cmp::Ordering::Equal)
+    // });
+
+    // table_rows
 }
 
 /// Summarizes portfolio data, calculating total value, weighted average percent change,
@@ -69,7 +99,7 @@ pub fn summarize_portfolio(table_rows: &[TableRow]) -> (f64, f64, f64, f64) {
     let weighted_average_percent_change: f64 = if total_value > 0.0 {
         table_rows
             .iter()
-            .map(|row| row.percent_change * (row.value.unwrap_or(0.0) / total_value))
+            .map(|row| row.hourly_percent_change * (row.value.unwrap_or(0.0) / total_value))
             .sum()
     } else {
         0.0
@@ -97,4 +127,22 @@ pub fn summarize_portfolio(table_rows: &[TableRow]) -> (f64, f64, f64, f64) {
         cumulative_pl,
         cumulative_pl_percentage,
     )
+}
+
+pub fn create_summary_row(table_rows: &Vec<TableRow>) -> TableRow {
+    let (total_value, weighted_average_percent_change, cumulative_pl, cumulative_pl_percentage) =
+        summarize_portfolio(table_rows);
+
+    TableRow {
+        ticker: "Summary".to_string(),
+        price: None,
+        hourly_percent_change: 0.0,
+        daily_percent_change: weighted_average_percent_change,
+        weekly_percent_change: 0.0,
+        entry_price: None,
+        amount: None,
+        value: Some(total_value),
+        pl: Some(cumulative_pl),
+        pl_percent: Some(cumulative_pl_percentage),
+    }
 }